+++ src/bract/ring/wrapper.clj
@@ -200,3 +200,41 @@
         (if-let [request (matcher request)]
           (handler request respond raise)
           (respond res-400))))))
+
+
+(defn wrap-params-normalize-wrapper
+  "Normalize the result of `wrap-params` middleware (this middleware may be invoked only after `wrap-params`) by
+  transforming each request params value. The `wrap-params` middleware extracts params as string, but multiple values
+  for same param are turned into a vector of string - this middleware turns all param values into vectors of string
+  and applies normalizer to that vector.
+
+  Without this middleware:
+  {\"foo\" \"bar\"
+   \"baz\" [\"quux\" \"corge\"]}
+
+  After using this middleware, where normalizer is `clojure.core/first`:
+  {\"foo\" \"bar\"
+   \"baz\" \"quux\"}"
+  [handler context normalizer]
+  (let [normalize (comp (core-type/ifunc normalizer) (fn [x] (cond
+                                                               (vector? x) x
+                                                               (coll? x)   (vec x)
+                                                               :otherwise  [x])))
+        transform (fn [m k]
+                    (if-let [pairs (get m k)]
+                      (->> pairs
+                        (reduce-kv (fn [mm kk vv] (assoc mm kk (normalize vv))) {})
+                        (assoc m k))
+                      m))]
+    (fn
+      ([request]
+        (handler (-> request
+                   (transform :form-params)
+                   (transform :query-params)
+                   (transform :params))))
+      ([request respond raise]
+        (handler (-> request
+                   (transform :form-params)
+                   (transform :query-params)
+                   (transform :params))
+          respond raise)))))